Feat/#201-K: 가입된 워크스페이스 없어도 워크스페이스 페이지로 이동 (#204)
