+++ client/src/App.tsx
@@ -15,7 +15,8 @@ function App() {
   const location = useLocation();
 
   const autoLogin = async () => {
-    const { user, workspaces } = await getAuth();
+    const authorizedUser = await getAuth();
+    const { user, workspaces } = authorizedUser;
 
     setIsLoaded(true);
 
@@ -24,9 +25,16 @@ function App() {
       return;
     }
 
-    setUserInfo({ user, workspaces });
+    setUserInfo(authorizedUser);
 
-    const { id } = workspaces[0];
+    if (!workspaces.length) {
+      navigate('/workspace');
+      return;
+    }
+
+    const defaultWorkspace = workspaces[0];
+    const { id } = defaultWorkspace;
+
     navigate(`/workspace/${id}`);
   };
 
@@ -39,7 +47,7 @@ function App() {
       <Routes>
         <Route path="/" element={<LoginPage />} />
         <Route path="/oauth" element={<OAuthPage />} />
-        <Route path="/workspace/:id" element={<WorkspacePage />} />
+        <Route path="/workspace/*" element={<WorkspacePage />} />
       </Routes>
     </UserContext.Provider>
   ) : (
